Suggestions from Deepak. Use xncml 0.3 from_text
