+++ STACpopulator/stac_utils.py
@@ -1,4 +1,5 @@
 import re
+from enum import Enum
 
 
 def url_validate(target: str) -> bool:
@@ -23,3 +24,20 @@ def url_validate(target: str) -> bool:
         re.IGNORECASE,
     )
     return True if re.match(url_regex, target) else False
+
+
+def collection2enum(collection):
+    """Create Enum based on terms from pyessv collection.
+
+    Parameters
+    ----------
+    collection : pyessv.model.collection.Collection
+      pyessv collection of terms.
+
+    Returns
+    -------
+    Enum
+      Enum storing terms and their labels from collection.
+    """
+    mp = {term.name: term.label for term in collection}
+    return Enum(collection.raw_name.capitalize(), mp, module="base")
