+++ pyproject.toml
@@ -11,6 +11,7 @@ dependencies = [
     "pyyaml",
     "siphon",
     "pystac",
+    "xncml",
     "pydantic",
     "pyessv"
 ]
